Wire Frame from MockFlow
This is a good point to start from design wise. The color Pallette hex values are as follows: #8BCDD0, #F2EDDF, #97B082, #7CA2A0, #618985 
This is subject to change as far as the overall design goes, but I believe the Header is a pretty solid figure. Also, the general layout is what she is looking for. Please feel free to use MockFlow html to adjust things or suggest improvements.